Move payload build into Slack class
